Initial attempt at adding Maven support. Moved stuff out of node support that wasn't directly tied to node. Use goal implementation push tests to differentiate node and maven goals. Made executeReleaseVersion generic rather than node specific by allowing an incrementPatchCmd to be specified.
