+++ package.json
@@ -31,6 +31,7 @@
     "@atomist/sdm-pack-changelog": "1.0.0-master.20180829151552",
     "@atomist/sdm-pack-k8": "1.0.0-master.20180829090607",
     "@atomist/sdm-pack-node": "1.0.0-master.20180829151541",
+    "@atomist/sdm-pack-spring": "^1.0.0-master.20180829151558",
     "@atomist/slack-messages": "^0.12.1",
     "fs-extra": "^7.0.0",
     "license-checker": "^20.2.0",
