+++ src/machine/nodeSupport.ts
@@ -14,39 +14,16 @@
  * limitations under the License.
  */
 
-import {
-    Configuration,
-    logger,
-} from "@atomist/automation-client";
 import { GitHubRepoRef } from "@atomist/automation-client/operations/common/GitHubRepoRef";
-import { GitProject } from "@atomist/automation-client/project/git/GitProject";
 import {
     allSatisfied,
     BuildGoal,
     hasFile,
     not,
-    ProductionEnvironment,
-    RepoContext,
-    SdmGoalEvent,
-    SoftwareDeliveryMachine,
-    SoftwareDeliveryMachineConfiguration,
-    StagingEnvironment,
 } from "@atomist/sdm";
-import {
-    createKubernetesData,
-    DefaultDockerImageNameCreator,
-    DockerBuildGoal,
-    DockerOptions,
-    executeDockerBuild,
-    executeTag,
-    executeVersioner,
-    TagGoal,
-    tagRepo,
-    VersionGoal,
-} from "@atomist/sdm-core";
-import { KubernetesOptions } from "@atomist/sdm-core/handlers/events/delivery/goals/k8s/launchGoalK8";
-import { changelogSupport } from "@atomist/sdm-pack-changelog";
-import { kubernetesSupport } from "@atomist/sdm-pack-k8";
+import {executeVersioner} from "@atomist/sdm-core/internal/delivery/build/local/projectVersioner";
+import {VersionGoal} from "@atomist/sdm-core/pack/well-known-goals/commonGoals";
+import {tagRepo} from "@atomist/sdm-core/util/github/tagRepo";
 import {
     executePublish,
     IsNode,
@@ -61,6 +38,7 @@ import {
 } from "@atomist/sdm-pack-node";
 import { executeBuild } from "@atomist/sdm/api-helper/goal/executeBuild";
 import { LogSuppressor } from "@atomist/sdm/api-helper/log/logInterpreters";
+import {SoftwareDeliveryMachine} from "@atomist/sdm/api/machine/SoftwareDeliveryMachine";
 import { AddAtomistTypeScriptHeader } from "../autofix/addAtomistHeader";
 import { AddThirdPartyLicense } from "../autofix/license/thirdPartyLicense";
 import { npmDockerfileFix } from "../autofix/npm/dockerfileFix";
@@ -71,24 +49,16 @@ import { TryToUpdateAtomistPeerDependencies } from "../transform/tryToUpdateAtom
 import { UpdatePackageAuthor } from "../transform/updatePackageAuthor";
 import { UpdatePackageVersion } from "../transform/updatePackageVersion";
 import {
-    ProductionDeploymentGoal,
     PublishGoal,
-    ReleaseChangelogGoal,
-    ReleaseDockerGoal,
     ReleaseDocsGoal,
     ReleaseNpmGoal,
-    ReleaseTagGoal,
     ReleaseVersionGoal,
     SmokeTestGoal,
-    StagingDeploymentGoal,
 } from "./goals";
 import {
-    DockerReleasePreparations,
     DocsReleasePreparations,
-    executeReleaseDocker,
     executeReleaseDocs,
     executeReleaseNpm,
-    executeReleaseTag,
     executeReleaseVersion,
     NpmReleasePreparations,
 } from "./release";
@@ -134,21 +104,6 @@ export function addNodeSupport(sdm: SoftwareDeliveryMachine): SoftwareDeliveryMa
             NodeDefaultOptions,
     )
         .addGoalImplementation(
-            "nodeDockerBuild",
-            DockerBuildGoal,
-            executeDockerBuild(
-                sdm.configuration.sdm.projectLoader,
-                DefaultDockerImageNameCreator,
-                NpmPreparations,
-                {
-                    ...sdm.configuration.sdm.docker.hub as DockerOptions,
-                    dockerfileFinder: async () => "Dockerfile",
-                }),
-            {
-                ...NodeDefaultOptions,
-            },
-    )
-        .addGoalImplementation(
             "nodePublish",
             PublishGoal,
             executePublish(sdm.configuration.sdm.projectLoader,
@@ -183,25 +138,6 @@ export function addNodeSupport(sdm: SoftwareDeliveryMachine): SoftwareDeliveryMa
             NodeDefaultOptions,
     )
         .addGoalImplementation(
-            "nodeDockerRelease",
-            ReleaseDockerGoal,
-            executeReleaseDocker(sdm.configuration.sdm.projectLoader,
-                DockerReleasePreparations,
-                {
-                    ...sdm.configuration.sdm.docker.hub as DockerOptions,
-                }),
-            {
-                pushTest: allSatisfied(IsNode, hasFile("Dockerfile")),
-                logInterpreter: NodeDefaultOptions.logInterpreter,
-            },
-    )
-        .addGoalImplementation(
-            "nodeTagRelease",
-            ReleaseTagGoal,
-            executeReleaseTag(sdm.configuration.sdm.projectLoader),
-            NodeDefaultOptions,
-    )
-        .addGoalImplementation(
             "nodeDocsRelease",
             ReleaseDocsGoal,
             executeReleaseDocs(sdm.configuration.sdm.projectLoader, DocsReleasePreparations),
@@ -212,27 +148,6 @@ export function addNodeSupport(sdm: SoftwareDeliveryMachine): SoftwareDeliveryMa
             ReleaseVersionGoal,
             executeReleaseVersion(sdm.configuration.sdm.projectLoader, NodeProjectIdentifier),
             NodeDefaultOptions,
-    )
-        .addGoalImplementation(
-            "nodeTag",
-            TagGoal,
-            executeTag(sdm.configuration.sdm.projectLoader),
-            NodeDefaultOptions,
-    );
-
-    sdm.addExtensionPacks(
-        changelogSupport(ReleaseChangelogGoal),
-        kubernetesSupport({
-            deployments: [{
-                goal: StagingDeploymentGoal,
-                pushTest: IsNode,
-                callback: kubernetesDataCallback(sdm.configuration),
-            }, {
-                goal: ProductionDeploymentGoal,
-                pushTest: IsNode,
-                callback: kubernetesDataCallback(sdm.configuration),
-            }],
-        }),
     );
 
     sdm.addAutofix(AddAtomistTypeScriptHeader)
@@ -253,86 +168,4 @@ export function addNodeSupport(sdm: SoftwareDeliveryMachine): SoftwareDeliveryMa
         .addCodeTransformCommand(UpdatePackageAuthor);
 
     return sdm;
-}
-
-function kubernetesDataCallback(
-    configuration: SoftwareDeliveryMachineConfiguration,
-): (goal: SdmGoalEvent, context: RepoContext) => Promise<SdmGoalEvent> {
-
-    return async (goal, ctx) => {
-        return configuration.sdm.projectLoader.doWithProject({
-            credentials: ctx.credentials, id: ctx.id, context: ctx.context, readOnly: true,
-        }, async p => {
-            return kubernetesDataFromGoal(goal, p, configuration);
-        });
-    };
-}
-
-function kubernetesDataFromGoal(
-    goal: SdmGoalEvent,
-    p: GitProject,
-    configuration: Configuration,
-): Promise<SdmGoalEvent> {
-
-    const ns = namespaceFromGoal(goal);
-    const ingress = ingressFromGoal(goal.repo.name, ns);
-    return createKubernetesData(
-        goal,
-        {
-            name: goal.repo.name,
-            environment: configuration.environment.split("_")[0],
-            port: 2866,
-            ns,
-            imagePullSecret: "atomistjfrog",
-            replicas: ns === "production" ? 3 : 1,
-            ...ingress,
-        } as KubernetesOptions,
-        p);
-}
-
-function namespaceFromGoal(goal: SdmGoalEvent): string {
-    const name = goal.repo.name;
-    if (/-sdm$/.test(name) && name !== "sample-sdm" && name !== "spring-sdm") {
-        return "sdm";
-    } else if (name === "k8-automation") {
-        return "k8-automation";
-    } else if (goal.environment === StagingEnvironment.replace(/\/$/, "")) {
-        return "testing";
-    } else if (goal.environment === ProductionEnvironment.replace(/\/$/, "")) {
-        return "production";
-    } else {
-        logger.debug(`Unmatched goal.environment using default namespace: ${goal.environment}`);
-        return "default";
-    }
-}
-
-export interface Ingress {
-    host: string;
-    path: string;
-    tlsSecret?: string;
-}
-
-export function ingressFromGoal(repo: string, ns: string): Ingress {
-    let host: string;
-    let path: string;
-    if (repo === "card-automation") {
-        host = "pusher";
-        path = "/";
-    } else if (repo === "sdm-automation") {
-        return {
-            host: "badge",
-            path: "/",
-        };
-    } else if (repo === "intercom-automation") {
-        host = "intercom";
-        path = "/";
-    } else {
-        return undefined;
-    }
-    const tail = (ns === "production") ? "com" : "services";
-    return {
-        host: `${host}.atomist.${tail}`,
-        path,
-        tlsSecret: `star-atomist-${tail}`,
-    };
 }
